+++ src/app/page.tsx
@@ -1,17 +1,15 @@
-"use client";
+import { Footer } from "@/components/common/Footer";
+import { Navbar } from "@/components/common/Navigationbar";
+import { SearchBar } from "@/components/home/SearchBar";
 
-import { Footer } from "../shared/components/common/Footer";
-import { Navbar } from "../shared/components/common/Navigationbar";
-import { SearchBar } from "../shared/components/home/SearchBar";
-
-export default function Home() {
+export default async function Home() {
 	return (
-		<div className="flex flex-col h-screen">
-			<Navbar />
-			<div className="flex-grow flex flex-col items-center justify-center">
-				<SearchBar />
+			<div className="flex flex-col h-screen p-0 m-0">
+				<Navbar/>
+				<div className="flex-grow flex flex-col items-center justify-center">
+					<SearchBar/>
+				</div>
+				<Footer/>
 			</div>
-			<Footer />
-		</div>
 	);
 }
