+++ src/components/__test__/common/footer.test.tsx
