+++ src/components/__test__/common/navigationBar.test.tsx
