+++ src/components/__test__/home/searchBar.test.tsx
